+++ algo/sword/search_53i.go
@@ -27,18 +27,3 @@ func search(nums []int, target int) int {
 
 	return count
 }
-
-// 查找在有序列表nums[first,stop)中第一个可插入位置
-func lowerBound(nums []int, value int) int {
-	left := 0
-	right := len(nums)
-	for left < right {
-		mid := left + (right-left)/2
-		if nums[mid] < value {
-			left = mid + 1
-		} else {
-			right = mid
-		}
-	}
-	return left
-}
